fix: restore style compliance and fix MavlinkCommandSender::Result format error
1. mavlink_component_metadata.cpp: use static_cast<int>(result) for
   MavlinkCommandSender::Result which has no operator<< and therefore
   no std::formatter specialization (fixes all-platform build failure).

2. log.h: apply clang-format-19 formatting:
   - #  define → #define (IndentPPDirectives: None)
   - requires clauses wrapped with opening paren on its own line
   - log_message fallback signature collapsed to one line (<100 chars)

3. mavlink_parameter_client.cpp: collapse 3-line LogDebug call to one
   line (fits within 100-char column limit).

4. mavsdk_server service_impl.h (38 files): restore static_cast<int>()
   in LogErr unknown-enum-value calls that were removed in the previous
   commit. These protobuf (unscoped) enum types lack operator<< and
   std::formatter, so direct {} formatting would fail to compile; the
   cast also matches what the Jinja2 template generates, keeping the
   style check clean.

+++ cpp/src/mavsdk_server/src/plugins/action/action_service_impl.h
@@ -91,7 +91,7 @@ public:
     {
         switch (relay_command) {
             default:
-                LogErr("Unknown relay_command enum value: {}", relay_command);
+                LogErr("Unknown relay_command enum value: {}", static_cast<int>(relay_command));
             // FALLTHROUGH
             case mavsdk::Action::RelayCommand::On:
                 return rpc::action::RELAY_COMMAND_ON;
@@ -105,7 +105,7 @@ public:
     {
         switch (relay_command) {
             default:
-                LogErr("Unknown relay_command enum value: {}", relay_command);
+                LogErr("Unknown relay_command enum value: {}", static_cast<int>(relay_command));
             // FALLTHROUGH
             case rpc::action::RELAY_COMMAND_ON:
                 return mavsdk::Action::RelayCommand::On;
@@ -119,7 +119,7 @@ public:
     {
         switch (result) {
             default:
-                LogErr("Unknown result enum value: {}", result);
+                LogErr("Unknown result enum value: {}", static_cast<int>(result));
             // FALLTHROUGH
             case mavsdk::Action::Result::Unknown:
                 return rpc::action::ActionResult_Result_RESULT_UNKNOWN;
@@ -159,7 +159,7 @@ public:
     {
         switch (result) {
             default:
-                LogErr("Unknown result enum value: {}", result);
+                LogErr("Unknown result enum value: {}", static_cast<int>(result));
             // FALLTHROUGH
             case rpc::action::ActionResult_Result_RESULT_UNKNOWN:
                 return mavsdk::Action::Result::Unknown;
