+++ cpp/src/mavsdk_server/src/plugins/mission/mission_service_impl.h
@@ -47,7 +47,7 @@ public:
     {
         switch (camera_action) {
             default:
-                LogErr("Unknown camera_action enum value: {}", camera_action);
+                LogErr("Unknown camera_action enum value: {}", static_cast<int>(camera_action));
             // FALLTHROUGH
             case mavsdk::Mission::MissionItem::CameraAction::None:
                 return rpc::mission::MissionItem_CameraAction_CAMERA_ACTION_NONE;
@@ -73,7 +73,7 @@ public:
     {
         switch (camera_action) {
             default:
-                LogErr("Unknown camera_action enum value: {}", camera_action);
+                LogErr("Unknown camera_action enum value: {}", static_cast<int>(camera_action));
             // FALLTHROUGH
             case rpc::mission::MissionItem_CameraAction_CAMERA_ACTION_NONE:
                 return mavsdk::Mission::MissionItem::CameraAction::None;
@@ -99,7 +99,7 @@ public:
     {
         switch (vehicle_action) {
             default:
-                LogErr("Unknown vehicle_action enum value: {}", vehicle_action);
+                LogErr("Unknown vehicle_action enum value: {}", static_cast<int>(vehicle_action));
             // FALLTHROUGH
             case mavsdk::Mission::MissionItem::VehicleAction::None:
                 return rpc::mission::MissionItem_VehicleAction_VEHICLE_ACTION_NONE;
@@ -119,7 +119,7 @@ public:
     {
         switch (vehicle_action) {
             default:
-                LogErr("Unknown vehicle_action enum value: {}", vehicle_action);
+                LogErr("Unknown vehicle_action enum value: {}", static_cast<int>(vehicle_action));
             // FALLTHROUGH
             case rpc::mission::MissionItem_VehicleAction_VEHICLE_ACTION_NONE:
                 return mavsdk::Mission::MissionItem::VehicleAction::None;
@@ -261,7 +261,7 @@ public:
     {
         switch (result) {
             default:
-                LogErr("Unknown result enum value: {}", result);
+                LogErr("Unknown result enum value: {}", static_cast<int>(result));
             // FALLTHROUGH
             case mavsdk::Mission::Result::Unknown:
                 return rpc::mission::MissionResult_Result_RESULT_UNKNOWN;
@@ -303,7 +303,7 @@ public:
     {
         switch (result) {
             default:
-                LogErr("Unknown result enum value: {}", result);
+                LogErr("Unknown result enum value: {}", static_cast<int>(result));
             // FALLTHROUGH
             case rpc::mission::MissionResult_Result_RESULT_UNKNOWN:
                 return mavsdk::Mission::Result::Unknown;
