+++ cpp/src/mavsdk_server/src/plugins/mission_raw/mission_raw_service_impl.h
@@ -185,7 +185,7 @@ public:
     {
         switch (result) {
             default:
-                LogErr("Unknown result enum value: {}", result);
+                LogErr("Unknown result enum value: {}", static_cast<int>(result));
             // FALLTHROUGH
             case mavsdk::MissionRaw::Result::Unknown:
                 return rpc::mission_raw::MissionRawResult_Result_RESULT_UNKNOWN;
@@ -239,7 +239,7 @@ public:
     {
         switch (result) {
             default:
-                LogErr("Unknown result enum value: {}", result);
+                LogErr("Unknown result enum value: {}", static_cast<int>(result));
             // FALLTHROUGH
             case rpc::mission_raw::MissionRawResult_Result_RESULT_UNKNOWN:
                 return mavsdk::MissionRaw::Result::Unknown;
